+++ controller/main.go
@@ -20,8 +20,8 @@ var assets embed.FS
 
 // Version information - injected at build time via -ldflags -X
 var (
-	Version     = "v3.1.121"
-	BuildDate   = "2026-08-20"
+	Version     = "v3.1.123"
+	BuildDate   = "2026-08-22"
 	VersionInfo = ""
 )